+++ src/main/java/chess/domain/piece/mulitiplemovepiece/MovingMultipleSquarePiece.java
@@ -1,15 +1,16 @@
-package chess.domain.piece;
+package chess.domain.piece.mulitiplemovepiece;
 
 import java.util.List;
 
+import chess.domain.piece.Color;
+import chess.domain.piece.Piece;
 import chess.domain.position.Direction;
 import chess.domain.position.UnitDirection;
 
-public abstract class MovingUnitSquarePiece extends Piece {
-
+public abstract class MovingMultipleSquarePiece extends Piece {
 	private final List<UnitDirection> directions;
 
-	MovingUnitSquarePiece(Color color, double score, List<UnitDirection> directions) {
+	MovingMultipleSquarePiece(Color color, double score, List<UnitDirection> directions) {
 		super(color, score);
 		this.directions = directions;
 	}
@@ -17,6 +18,6 @@ public abstract class MovingUnitSquarePiece extends Piece {
 	@Override
 	public final boolean canMove(Direction direction, Piece target) {
 		checkSameTeam(target);
-		return direction.hasSame(directions);
+		return direction.hasMultiple(directions);
 	}
 }
