+++ src/main/java/chess/domain/piece/unitmovepiece/MovingUnitSquarePiece.java
@@ -1,14 +1,17 @@
-package chess.domain.piece;
+package chess.domain.piece.unitmovepiece;
 
 import java.util.List;
 
+import chess.domain.piece.Color;
+import chess.domain.piece.Piece;
 import chess.domain.position.Direction;
 import chess.domain.position.UnitDirection;
 
-public abstract class MovingMultipleSquarePiece extends Piece {
+public abstract class MovingUnitSquarePiece extends Piece {
+
 	private final List<UnitDirection> directions;
 
-	MovingMultipleSquarePiece(Color color, double score, List<UnitDirection> directions) {
+	MovingUnitSquarePiece(Color color, double score, List<UnitDirection> directions) {
 		super(color, score);
 		this.directions = directions;
 	}
@@ -16,6 +19,6 @@ public abstract class MovingMultipleSquarePiece extends Piece {
 	@Override
 	public final boolean canMove(Direction direction, Piece target) {
 		checkSameTeam(target);
-		return direction.hasMultiple(directions);
+		return direction.hasSame(directions);
 	}
 }
